Fixed type error of range parameter
n_pos was of type float but range wants integer.

Error experienced with the files named 314_RZ_9847886_Jun-06-14-05-27
for the game alien from the Atari-Head dataset (https://zenodo.org/record/2603190).

+++ data_reader.py
@@ -203,7 +203,7 @@ def read_gaze_data_csv_file(fname, separator=',', pos_separator=','):
             else:
                 pos_data_list = pos_data.split(pos_separator)
             pos_list = []
-            n_pos = len(pos_data_list) / 2
+            n_pos = int(len(pos_data_list) / 2)
             for j in range(0, n_pos):
                 posX = float(pos_data_list[2*j])
                 posY = float(pos_data_list[2*j+1])
